update oa question operator and student info add

+++ src/main/java/com/duyi/oa/domain/QuestionBody.java
@@ -2,8 +2,8 @@ package com.duyi.oa.domain;
 
 public class QuestionBody {
     private int id = 0;
-    private int uploadId;
-    private  int reviewerId;
+    private Long uploadId;
+    private  Long reviewerId;
     private int topicId;
     private int status;
     private String title;
@@ -21,10 +21,8 @@ public class QuestionBody {
         this.status = status;
     }
 
-    public  int getUploadId(){ return  this.uploadId; }
-    public void setUploadId(int uploadId){
-        this.uploadId = uploadId;
-    }
+    public Long getUploadId(){ return  this.uploadId; }
+    public void setUploadId(Long uploadId){ this.uploadId = uploadId; }
 
     public int getTopicdId(){
         return  this.topicId;
@@ -55,6 +53,6 @@ public class QuestionBody {
     public String getPic(){ return  this.pic; }
     public  void setPic(String pic){ this.pic = pic; }
 
-    public int getReviewerId(){ return  this.reviewerId;}
-    public void setReviewerId(int reviewerId){ this.reviewerId = reviewerId; }
+    public Long getReviewerId(){ return  this.reviewerId;}
+    public void setReviewerId(Long reviewerId){ this.reviewerId = reviewerId; }
 }
